Implement points multiplication
The used method offers a very small increase in performance at the cost
of readability over the other method which simply loops trough the list
twice, once to find the exams with the most points and once to add up
the points of all the other exams.
Also this method is longer, more complex, harder to maintain and
took longer to write.
Yes, I should have used the other method.
No, I am not going to use the other method.