Remove save/discard system
The system did not actually cache changes, it just kept a copy of
the old data and set the actual values to the previously made copy if
necessary. This interfered with the database because changes would be
sent to the database before the save button was pressed, which is
undesired behaviour.